// packages/headless/src/modules/dataManager/types.ts
// --- external
import type { JsonSchema, UISchemaElement } from "@jsonforms/core";
import { BrandConfigKeys, type IClient } from "@upmind-automation/types";

// --- internal
import type { ResponseError } from "../../utils";

// -----------------------------------------------------------------------------
// Contexts

export interface DataManagerContext<TModel = any, TBaseModel = any> {
  clientId?: IClient["id"]; // allow for an override of the current client
  id?: string;

  title?: string;
  description?: string;
  lookups?: Record<string, any[]>;
  // ---
  schema?: JsonSchema;
  uischema?: UISchemaElement;
  baseModel?: TBaseModel;
  model?: TModel;
  error?: ResponseError;
  // ---
  autoupdate?: boolean;
  allowMultipleEdits?: boolean;
  // ---
  config?: Record<BrandConfigKeys, boolean>;
}